requirements.txt
telethon>=1.25.0
pymongo>=4.0
motor>=2.5
aiohttp>=3.8.0
asyncio
elasticsearch>=8.0
redis>=4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
cryptography>=40.0.0
pytz
dnspython
pymongo[srv]
